backend/requirements.txt
flask
Flask-SQLAlchemy 
Flask-Bcrypt 
PyJWT
Flask-JWT-Extended
flask-cors
fastapi[all]
fastapi
uvicorn
markdown
google-api-python-client
asyncio
deepgram-sdk
groq
ipykernel
langchain
langchain-community
langchain-core
langchain-groq
langchain-openai
langsmith
numpy
openai
PyAudio
python-dotenv
requests
websocket
websocket-client
sounddevice
Flask-Mail